fix extension dependencies
