+++ bom/drools-bom/pom.xml
@@ -438,36 +438,36 @@
       <!-- Quarkus Extension -->
       <dependency>
         <groupId>org.drools</groupId>
-        <artifactId>drools-quarkus</artifactId>
+        <artifactId>drools-drl-quarkus</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
         <groupId>org.drools</groupId>
-        <artifactId>drools-quarkus</artifactId>
+        <artifactId>drools-drl-quarkus</artifactId>
         <version>${project.version}</version>
         <classifier>sources</classifier>
       </dependency>
       <dependency>
         <groupId>org.drools</groupId>
-        <artifactId>drools-quarkus</artifactId>
+        <artifactId>drools-drl-quarkus</artifactId>
         <version>${project.version}</version>
         <type>test-jar</type>
         <scope>test</scope>
       </dependency>
       <dependency>
         <groupId>org.drools</groupId>
-        <artifactId>drools-quarkus-deployment</artifactId>
+        <artifactId>drools-drl-quarkus-deployment</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
         <groupId>org.drools</groupId>
-        <artifactId>drools-quarkus-deployment</artifactId>
+        <artifactId>drools-drl-quarkus-deployment</artifactId>
         <version>${project.version}</version>
         <classifier>sources</classifier>
       </dependency>
       <dependency>
         <groupId>org.drools</groupId>
-        <artifactId>drools-quarkus-deployment</artifactId>
+        <artifactId>drools-drl-quarkus-deployment</artifactId>
         <version>${project.version}</version>
         <type>test-jar</type>
         <scope>test</scope>
