+++ drools-drl-quarkus-extension/drools-drl-quarkus-deployment/src/main/java/org/drools/drl/quarkus/deployment/DroolsAssetsProcessor.java
@@ -1,0 +1,97 @@
+/*
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.drools.drl.quarkus.deployment;
+
+import java.util.Collection;
+import javax.inject.Inject;
+
+import io.quarkus.arc.deployment.GeneratedBeanBuildItem;
+import io.quarkus.deployment.annotations.BuildProducer;
+import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.builditem.ArchiveRootBuildItem;
+import io.quarkus.deployment.builditem.CombinedIndexBuildItem;
+import io.quarkus.deployment.builditem.FeatureBuildItem;
+import io.quarkus.deployment.builditem.GeneratedResourceBuildItem;
+import io.quarkus.deployment.builditem.LiveReloadBuildItem;
+import io.quarkus.deployment.builditem.nativeimage.NativeImageResourceBuildItem;
+import io.quarkus.deployment.pkg.builditem.CurateOutcomeBuildItem;
+import io.quarkus.deployment.pkg.builditem.OutputTargetBuildItem;
+import io.quarkus.maven.dependency.ResolvedDependency;
+import io.quarkus.vertx.http.deployment.spi.AdditionalStaticResourceBuildItem;
+import org.drools.model.project.codegen.GeneratedFile;
+import org.drools.model.project.codegen.GeneratedFileType;
+import org.drools.model.project.codegen.context.DroolsModelBuildContext;
+import org.kie.api.io.Resource;
+
+import static org.drools.drl.quarkus.deployment.DroolsQuarkusResourceUtils.HOT_RELOAD_SUPPORT_PATH;
+import static org.drools.drl.quarkus.deployment.DroolsQuarkusResourceUtils.compileGeneratedSources;
+import static org.drools.drl.quarkus.deployment.DroolsQuarkusResourceUtils.createDroolsBuildContext;
+import static org.drools.drl.quarkus.deployment.DroolsQuarkusResourceUtils.dumpFilesToDisk;
+import static org.drools.drl.quarkus.deployment.DroolsQuarkusResourceUtils.getHotReloadSupportSource;
+import static org.drools.drl.quarkus.deployment.DroolsQuarkusResourceUtils.registerResources;
+import static org.drools.model.project.codegen.RuleCodegen.ofResources;
+
+public class DroolsAssetsProcessor {
+
+    @Inject
+    ArchiveRootBuildItem root;
+    @Inject
+    LiveReloadBuildItem liveReload;
+    @Inject
+    CurateOutcomeBuildItem curateOutcomeBuildItem;
+    @Inject
+    CombinedIndexBuildItem combinedIndexBuildItem;
+    @Inject
+    OutputTargetBuildItem outputTargetBuildItem;
+
+    private static final String FEATURE = "drools-drl-quarkus-extension";
+
+    @BuildStep
+    public FeatureBuildItem feature() {
+        return new FeatureBuildItem(FEATURE);
+    }
+
+    @BuildStep
+    public void generateSources( BuildProducer<GeneratedBeanBuildItem> generatedBeans,
+                                 BuildProducer<NativeImageResourceBuildItem> resource,
+                                 BuildProducer<AdditionalStaticResourceBuildItem> staticResProducer,
+                                 BuildProducer<GeneratedResourceBuildItem> genResBI ) {
+        DroolsModelBuildContext context =
+                createDroolsBuildContext(outputTargetBuildItem.getOutputDirectory(), root.getPaths(), combinedIndexBuildItem.getIndex());
+
+        Collection<Resource> resources = ResourceCollector.fromPaths(context.getAppPaths().getPaths());
+
+        Collection<GeneratedFile> generatedFiles = ofResources(context, resources).generate();
+
+        // The HotReloadSupportClass has to be generated only during the first model generation
+        // During actual hot reloads it will be regenerated by the compilation providers in order to retrigger this build step
+        if (!liveReload.isLiveReload()) {
+            generatedFiles.add(new GeneratedFile(GeneratedFileType.SOURCE, HOT_RELOAD_SUPPORT_PATH + ".java", getHotReloadSupportSource()));
+        }
+
+        // dump files to disk
+        dumpFilesToDisk(context.getAppPaths(), generatedFiles);
+
+        // build Java source code and register the generated beans
+        Collection<ResolvedDependency> dependencies = curateOutcomeBuildItem.getApplicationModel().getRuntimeDependencies();
+
+        Collection<GeneratedBeanBuildItem> generatedBeanBuildItems =
+                compileGeneratedSources(context, dependencies, generatedFiles, liveReload.isLiveReload());
+        generatedBeanBuildItems.forEach(generatedBeans::produce);
+
+        registerResources(generatedFiles, staticResProducer, resource, genResBI);
+    }
+}
