+++ drools-drl-quarkus-extension/drools-drl-quarkus-deployment/src/main/resources/META-INF/services/io.quarkus.deployment.dev.CompilationProvider
@@ -1,0 +1,1 @@
+org.drools.drl.quarkus.deployment.DroolsCompilationProvider
