+++ drools-drl-quarkus-extension/drools-drl-quarkus-integration-test-hotreload/src/main/resources/META-INF/kmodule.xml
