+++ drools-drl-quarkus-extension/drools-drl-quarkus-integration-test-hotreload/src/main/resources/application.properties
