+++ drools-drl-quarkus-extension/drools-drl-quarkus-integration-test-hotreload/src/main/resources/org/drools/hotreload/adult.txt
