+++ drools-drl-quarkus-extension/drools-drl-quarkus-integration-test/src/main/resources/META-INF/kmodule.xml
