+++ drools-drl-quarkus-extension/drools-drl-quarkus-integration-test/src/main/resources/org/drools/cep/cep.drl
