+++ drools-drl-quarkus-extension/drools-drl-quarkus-integration-test/src/main/resources/org/drools/probe/probe.drl
