+++ drools-drl-quarkus-extension/drools-drl-quarkus/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,7 @@
+---
+name: "Drools-DRL"
+metadata:
+  keywords:
+  - "drools"
+  - "rules"
+  - "DRL"
