+++ drools-drl-quarkus-extension/pom.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.kie</groupId>
+    <artifactId>drools-build-parent</artifactId>
+    <version>8.19.0-SNAPSHOT</version>
+    <relativePath>../build-parent/pom.xml</relativePath>
+  </parent>
+
+  <name>Drools :: Quarkus Extension</name>
+  <groupId>org.drools</groupId>
+  <artifactId>drools-drl-quarkus-extension</artifactId>
+  <packaging>pom</packaging>
+
+  <properties>
+    <version.org.apache.maven>3.8.4</version.org.apache.maven>
+  </properties>
+
+  <modules>
+    <module>drools-drl-quarkus-deployment</module>
+    <module>drools-drl-quarkus</module>
+    <module>drools-drl-quarkus-integration-test</module>
+    <module>drools-drl-quarkus-integration-test-hotreload</module>
+  </modules>
+
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>io.quarkus</groupId>
+        <artifactId>quarkus-bom</artifactId>
+        <version>${version.io.quarkus}</version>
+        <type>pom</type>
+        <scope>import</scope>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
+
+</project>
