--- drools-quarkus-extension/drools-quarkus-deployment/src/main/java/org/drools/quarkus/deployment/DroolsBuildContextBuildItem.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.drools.quarkus.deployment;
-
-import io.quarkus.builder.item.SimpleBuildItem;
-import org.drools.model.project.codegen.context.DroolsModelBuildContext;
-
-public final class DroolsBuildContextBuildItem extends SimpleBuildItem {
-
-    private final DroolsModelBuildContext droolsModelBuildContext;
-
-    public DroolsBuildContextBuildItem(DroolsModelBuildContext droolsModelBuildContext) {
-        this.droolsModelBuildContext = droolsModelBuildContext;
-    }
-
-    public DroolsModelBuildContext getDroolsModelBuildContext() {
-        return droolsModelBuildContext;
-    }
-}
