--- drools-quarkus-extension/drools-quarkus-deployment/src/main/resources/META-INF/services/io.quarkus.deployment.dev.CompilationProvider
@@ -1,1 +1,0 @@
-org.drools.quarkus.deployment.DroolsCompilationProvider
