--- drools-quarkus-extension/drools-quarkus/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,10 +1,0 @@
----
-name: "Drools"
-metadata:
-  keywords:
-  - "drools"
-  - "rules"
-  - "DRL"
-  guide: "https://quarkus.io/guides/kogito-drl"
-  categories:
-  - "business-automation"
